Refactor custom_ssh_port access helper method for consistency
Rather than repeat all the specifications around exceptions etc,
properly extract and refactor the custom_ssh_port method so that
it is always available in recipes and is directly unit tested.
