+++ .travis.yml
@@ -4,5 +4,5 @@ rvm:
   - 2.3.1
 sudo:     false
 script:
-  - bundle exec foodcritic -f any .
+  - bundle exec foodcritic -f any . -X spec
   - bundle exec rspec --color --format progress
