+++ spec/recipes/localhost_aliases_spec.rb
